+++ Pulumi.mjs
@@ -1,18 +1,20 @@
 import { createUploads } from './pulumi/uploads.mjs';
 import { createVpc } from './pulumi/vpc.mjs';
+import { createAurora } from './pulumi/aurora.mjs';
 import { createThumbnails } from './pulumi/thumbnails.mjs';
 import { createLambdaProcessUploads } from './pulumi/lambda-process-uploads.mjs';
 import { updateEnv } from './pulumi/env.mjs';
 import * as pulumi from "@pulumi/pulumi";
 
 const stack = pulumi.getStack();
-const { uploadsBucket, uploadsAccessKey } = createUploads();
-const thumbnailsBucket = createThumbnails();
-const processUploadsFunctionUrl = createLambdaProcessUploads(thumbnailsBucket);
+const { uploadsBucket, uploadsAccessKey } = await createUploads();
+const thumbnailsBucket = await createThumbnails();
+const processUploadsFunctionUrl = await createLambdaProcessUploads(thumbnailsBucket);
 if (stack != 'dev') {
-    const vpc = createVpc();
+    const { vpc, subnets } = await createVpc();
+    const { auroraCluster } = await createAurora(vpc, subnets)
 }
-updateEnv({
+await updateEnv({
     stack,
     uploadsBucket,
     uploadsAccessKey,