+++ pulumi/env.mjs
@@ -4,7 +4,7 @@ import * as pulumi from "@pulumi/pulumi";
 
 // this has to be its own file rather than each function updating itself, because of the paralleization of pulumi causing a race condition for the .env file
 
-export const updateEnv = (obj) => {
+export const updateEnv = async (obj) => {
     const env = new Env(obj.stack == 'dev' ? '.env' : '.env.' + obj.stack);
     aws.getRegion({}).then(region => env.set('AWS_REGION', region.id));
     obj.uploadsAccessKey.id.apply(id => env.set("VING_AWS_UPLOADS_KEY", id));