+++ pulumi/utils.mjs
@@ -1,5 +1,11 @@
 import * as pulumi from "@pulumi/pulumi";
 
+export const safeName = () => {
+    return `${pulumi.getProject()}${pulumi.getStack()}`.toLowerCase().replace(/\s+/g, '');
+}
+
 export const prefix = (name) => {
-    return `${pulumi.getProject()}-${pulumi.getStack()}-${name}`;
+    if (name)
+        return `${pulumi.getProject()}-${pulumi.getStack()}-${name}`;
+    return `${pulumi.getProject()}-${pulumi.getStack()}`;
 }