+++ pulumi/vpc.mjs
@@ -1,7 +1,7 @@
 import * as aws from "@pulumi/aws";
 import { prefix } from './utils.mjs';
 
-export const createVpc = () => {
+export const createVpc = async () => {
 
     const vpc = new aws.ec2.Vpc(prefix('vpc'), {
         cidrBlock: "10.0.0.0/16",
@@ -12,5 +12,19 @@ export const createVpc = () => {
         },
     });
 
-    return vpc;
+    const region = await aws.getRegion({});
+
+    const subnet1 = new aws.ec2.Subnet(prefix('subnet1'), {
+        vpcId: vpc.id,
+        cidrBlock: "10.0.1.0/24",
+        availabilityZone: region.id + "a",
+    });
+
+    const subnet2 = new aws.ec2.Subnet(prefix('subnet2'), {
+        vpcId: vpc.id,
+        cidrBlock: "10.0.2.0/24",
+        availabilityZone: region.id + "b",
+    });
+
+    return { vpc, subnets: [subnet1, subnet2] };
 }