+++ ving/docs/change-log.md
@@ -8,6 +8,7 @@ outline: deep
 ### 2024-08-02
 * Added Pulumi prod.
 * Pulumi prod creates a VPC.
+* Pulumi prod creates a a database cluster.
 
 ## July 2024